test(sdk): rebuild pytest suite for the platform x plugin QDC matrix
Consolidate every plugin's LLM/VLM/precision/MTP coverage into a single
per-plugin file (tests/test_llama_cpp.py, tests/test_qairt.py), each
holding four cases: multi-turn LLM, multi-turn VLM, keyword precision,
and MTP (llama_cpp only). Multi-turn subsumes the previous single-turn
smokes and the redundant ngram-simple case; MTP moves into the same
file since one QDC leg now runs both plugins.

Add an explicit test_model_manager_pull per plugin file. Fixture pull
errors bubble up as FAIL instead of pytest.skip so a broken AI Hub /
HuggingFace pull surfaces red, not silently green.

Flatten tests/api/ into tests/test_api.py — five ~20-line files did not
earn their directory. conftest.py's auto-tag derives the marker from
the filename stem now.

Delete tests/cli/ (no runner references it; its BUILD.bazel points at
a test.py that does not exist) and drop the device_gpu / device_hybrid
markers (neither backend is in the production matrix).
