--- tests/cli/.gitattributes
@@ -1,1 +1,0 @@
-assets/* filter=lfs diff=lfs merge=lfs -text
