--- tests/cli/BUILD.bazel
@@ -1,7 +1,0 @@
-load("@rules_python//python:defs.bzl", "py_binary")
-
-py_binary(
-    name = "cli",
-    srcs = ["test.py"],
-    main = "test.py",
-)
