--- tests/cli/assets/cat.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d3ba02173242b299b9aaf68386f0f28cf7434e13294e57beb65d90c28d0487c9
-size 695195
