--- tests/cli/assets/storytelling.wav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2812d093bdfe932f0d063d1ae194ad0ebb8b4be2bfab2dd0ab36c420a354b783
-size 785358
