--- tests/cli/assets/text.jpeg
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c33036eced12d529a3855c065019ae5102e48285c8116bdef1c64cc87d1165c9
-size 199851
