--- tests/cli/cases/__init__.py
@@ -1,16 +1,0 @@
-# Copyright 2024-2026 Qualcomm Technologies, Inc. and/or its subsidiaries.
-# SPDX-License-Identifier: BSD-3-Clause
-
-from .audio_multi_round import AudioMultiRound
-from .base import BaseCase
-from .image_multi_round import ImageMultiRound
-from .multi_round import MultiRound
-from .single_round import SingleRound
-
-__all__ = [
-    "BaseCase",
-    "SingleRound",
-    "MultiRound",
-    "ImageMultiRound",
-    "AudioMultiRound",
-]
