+++ tests/pytest.ini
@@ -9,8 +9,6 @@ markers =
     llm: text-only generation
     vlm: vision-language generation
     device_cpu: cpu backend (works anywhere)
-    device_gpu: GPUOpenCL backend (Adreno on Snapdragon)
     device_npu: HTP NPU backend (Snapdragon only)
-    device_hybrid: HTP+CPU hybrid scheduler (Snapdragon only)
     snapdragon: requires real Snapdragon host (gated by GENIEX_DEVICE_TEST=1)
-norecursedirs = cli qdc
+norecursedirs = qdc
